Testing JDBC

+++ src/main/java/br/com/juliocauan/familybudget/domain/application/model/Revenue.java
@@ -1,9 +1,10 @@
 package br.com.juliocauan.familybudget.domain.application.model;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 public interface Revenue {
     String getDescription();
-    Float getValue();
+    BigDecimal getValue();
     Date getDate();
 }
